refactor: input validator 추가
